// frontend/src/openapi/models/CapabilityFlow.ts
export interface CapabilityFlow {
    /**
     *
     * @type {boolean}
     * @memberof CapabilityFlow
     */
    hasCard: boolean;
    /**
     *
     * @type {boolean}
     * @memberof CapabilityFlow
     */
    isFallback: boolean;
    /**
     *
     * @type {string}
     * @memberof CapabilityFlow
     */
    source: string;
    /**
     *
     * @type {string}
     * @memberof CapabilityFlow
     */
    task: string;
    /**
     *
     * @type {string}
     * @memberof CapabilityFlow
     */
    modality: string;
    /**
     *
     * @type {string}
     * @memberof CapabilityFlow
     */
    domain: string;
    /**
     *
     * @type {string}
     * @memberof CapabilityFlow
     */
    model: string;
    /**
     *
     * @type {string}
     * @memberof CapabilityFlow
     */
    systemPrompt: string;
    /**
     *
     * @type {string}
     * @memberof CapabilityFlow
     */
    systemPromptExcerpt: string;
    /**
     *
     * @type {boolean}
     * @memberof CapabilityFlow
     */
    takesFiles: boolean;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof CapabilityFlow
     */
    inputSchema: { [key: string]: any; };
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof CapabilityFlow
     */
    outputFields: { [key: string]: any; };
    /**
     *
     * @type {CapabilityFlowExpectedOutput}
     * @memberof CapabilityFlow
     */
    expectedOutput: CapabilityFlowExpectedOutput;
    /**
     *
     * @type {Array<CapabilityFlowTool>}
     * @memberof CapabilityFlow
     */
    toolSpec: Array<CapabilityFlowTool>;
    /**
     *
     * @type {Array<CapabilityFlowMode>}
     * @memberof CapabilityFlow
     */
    modes: Array<CapabilityFlowMode>;
    /**
     *
     * @type {Array<CapabilityFlowUtility>}
     * @memberof CapabilityFlow
     */
    llmUtilities: Array<CapabilityFlowUtility>;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof CapabilityFlow
     */
    vocabulary: { [key: string]: any; };
    /**
     *
     * @type {Array<string>}
     * @memberof CapabilityFlow
     */
    successCriteria: Array<string>;
    /**
     *
     * @type {Array<string>}
     * @memberof CapabilityFlow
     */
    failureModes: Array<string>;
    /**
     *
     * @type {Array<CapabilityFlowTrajectoryPath>}
     * @memberof CapabilityFlow
     */
    trajectoryMap: Array<CapabilityFlowTrajectoryPath>;
    /**
     *
     * @type {Array<string>}
     * @memberof CapabilityFlow
     */
    provenancePaths: Array<string>;
}